+++ goclient/testdata/fixture.go_server_methods.go
@@ -22,7 +22,8 @@ import (
 )
 
 // Authenticate post /auth
-func (_c Client) Authenticate(ctx context.Context, baseURL BaseURLBuilder) (AuthenticateResponse, *http.Response, error) {
+func (_c Client) Authenticate(ctx context.Context) (AuthenticateResponse, *http.Response, error) {
+	baseURL := _c.url
 	_urlStr := strings.TrimSuffix(baseURL.ToURL(), "/") + `/auth`
 	_req, _err := http.NewRequestWithContext(ctx, http.MethodPost, _urlStr, nil)
 	if _err != nil {
@@ -50,7 +51,8 @@ func (_c Client) Authenticate(ctx context.Context, baseURL BaseURLBuilder) (Auth
 }
 
 // TestArrayRequest get /test/array/request
-func (_c Client) TestArrayRequest(ctx context.Context, baseURL BaseURLBuilder, body TestArrayRequestInline) (TestArrayRequestResponse, *http.Response, error) {
+func (_c Client) TestArrayRequest(ctx context.Context, body TestArrayRequestInline) (TestArrayRequestResponse, *http.Response, error) {
+	baseURL := _c.url
 	_urlStr := strings.TrimSuffix(baseURL.ToURL(), "/") + `/test/array/request`
 	_req, _err := http.NewRequestWithContext(ctx, http.MethodGet, _urlStr, nil)
 	if _err != nil {
@@ -83,7 +85,8 @@ func (_c Client) TestArrayRequest(ctx context.Context, baseURL BaseURLBuilder, b
 }
 
 // TestEnumQueryRequest get /test/enum/query/request
-func (_c Client) TestEnumQueryRequest(ctx context.Context, baseURL BaseURLBuilder, body TestEnumQueryRequestInline, sort TestEnumQueryRequestGetSortParam) (TestEnumQueryRequestResponse, *http.Response, error) {
+func (_c Client) TestEnumQueryRequest(ctx context.Context, body TestEnumQueryRequestInline, sort TestEnumQueryRequestGetSortParam) (TestEnumQueryRequestResponse, *http.Response, error) {
+	baseURL := _c.url
 	_urlStr := strings.TrimSuffix(baseURL.ToURL(), "/") + `/test/enum/query/request`
 	_req, _err := http.NewRequestWithContext(ctx, http.MethodGet, _urlStr, nil)
 	if _err != nil {
@@ -120,7 +123,8 @@ func (_c Client) TestEnumQueryRequest(ctx context.Context, baseURL BaseURLBuilde
 }
 
 // TestInlinePrimitiveBody get /test/inline
-func (_c Client) TestInlinePrimitiveBody(ctx context.Context, baseURL BaseURLBuilder, body string) (TestInlinePrimitiveBodyResponse, *http.Response, error) {
+func (_c Client) TestInlinePrimitiveBody(ctx context.Context, body string) (TestInlinePrimitiveBodyResponse, *http.Response, error) {
+	baseURL := _c.url
 	_urlStr := strings.TrimSuffix(baseURL.ToURL(), "/") + `/test/inline`
 	_req, _err := http.NewRequestWithContext(ctx, http.MethodGet, _urlStr, nil)
 	if _err != nil {
@@ -153,7 +157,8 @@ func (_c Client) TestInlinePrimitiveBody(ctx context.Context, baseURL BaseURLBui
 }
 
 // TestInline post /test/inline
-func (_c Client) TestInline(ctx context.Context, baseURL BaseURLBuilder, body TestInlineInline) (TestInlineResponse, *http.Response, error) {
+func (_c Client) TestInline(ctx context.Context, body TestInlineInline) (TestInlineResponse, *http.Response, error) {
+	baseURL := _c.url
 	_urlStr := strings.TrimSuffix(baseURL.ToURL(), "/") + `/test/inline`
 	_req, _err := http.NewRequestWithContext(ctx, http.MethodPost, _urlStr, nil)
 	if _err != nil {
@@ -204,7 +209,7 @@ func (_c Client) TestInline(ctx context.Context, baseURL BaseURLBuilder, body Te
 }
 
 // TestServerPathOverrideRequest get /test/servers
-func (_c Client) TestServerPathOverrideRequest(ctx context.Context, baseURL BaseURLBuilderTestservers) (TestServerPathOverrideRequestResponse, *http.Response, error) {
+func (_c Client) TestServerPathOverrideRequest(ctx context.Context, baseURL URLBuilderTestservers) (TestServerPathOverrideRequestResponse, *http.Response, error) {
 	_urlStr := strings.TrimSuffix(baseURL.ToURL(), "/") + `/test/servers`
 	_req, _err := http.NewRequestWithContext(ctx, http.MethodGet, _urlStr, nil)
 	if _err != nil {
@@ -232,7 +237,7 @@ func (_c Client) TestServerPathOverrideRequest(ctx context.Context, baseURL Base
 }
 
 // TestServerOpOverrideRequest post /test/servers
-func (_c Client) TestServerOpOverrideRequest(ctx context.Context, baseURL BaseURLBuilderTestserversPost) (TestServerOpOverrideRequestResponse, *http.Response, error) {
+func (_c Client) TestServerOpOverrideRequest(ctx context.Context, baseURL URLBuilderTestserversPost) (TestServerOpOverrideRequestResponse, *http.Response, error) {
 	_urlStr := strings.TrimSuffix(baseURL.ToURL(), "/") + `/test/servers`
 	_req, _err := http.NewRequestWithContext(ctx, http.MethodPost, _urlStr, nil)
 	if _err != nil {
@@ -259,8 +264,67 @@ func (_c Client) TestServerOpOverrideRequest(ctx context.Context, baseURL BaseUR
 	return _resp, _httpResp, nil
 }
 
+// TestSingleServerPathOverrideRequest get /test/single_servers
+func (_c Client) TestSingleServerPathOverrideRequest(ctx context.Context) (TestSingleServerPathOverrideRequestResponse, *http.Response, error) {
+	baseURL := Httppathdevlocal
+	_urlStr := strings.TrimSuffix(baseURL.ToURL(), "/") + `/test/single_servers`
+	_req, _err := http.NewRequestWithContext(ctx, http.MethodGet, _urlStr, nil)
+	if _err != nil {
+		return nil, nil, _err
+	}
+	var _query url.Values
+	if len(_query) > 0 {
+		_req.URL.RawQuery = _query.Encode()
+	}
+
+	_httpResp, _err := _c.doRequest(ctx, _req)
+	if _err != nil {
+		return nil, nil, _err
+	}
+
+	var _resp TestSingleServerPathOverrideRequestResponse
+	switch _httpResp.StatusCode {
+	case 200:
+		_resp = HTTPStatusOk{}
+	default:
+		return nil, _httpResp, fmt.Errorf("unknown response code %d", _httpResp.StatusCode)
+	}
+
+	return _resp, _httpResp, nil
+}
+
+// TestSingleServerOpOverrideRequest post /test/single_servers
+func (_c Client) TestSingleServerOpOverrideRequest(ctx context.Context) (TestSingleServerOpOverrideRequestResponse, *http.Response, error) {
+	baseURL := Httpopdevlocal
+	_urlStr := strings.TrimSuffix(baseURL.ToURL(), "/") + `/test/single_servers`
+	_req, _err := http.NewRequestWithContext(ctx, http.MethodPost, _urlStr, nil)
+	if _err != nil {
+		return nil, nil, _err
+	}
+	var _query url.Values
+	if len(_query) > 0 {
+		_req.URL.RawQuery = _query.Encode()
+	}
+
+	_httpResp, _err := _c.doRequest(ctx, _req)
+	if _err != nil {
+		return nil, nil, _err
+	}
+
+	var _resp TestSingleServerOpOverrideRequestResponse
+	switch _httpResp.StatusCode {
+	case 200:
+		_resp = HTTPStatusOk{}
+	default:
+		return nil, _httpResp, fmt.Errorf("unknown response code %d", _httpResp.StatusCode)
+	}
+
+	return _resp, _httpResp, nil
+}
+
 // TestTypeOverrides get /test/type_overrides
-func (_c Client) TestTypeOverrides(ctx context.Context, baseURL BaseURLBuilder, body *Primitives, number decimal.Decimal, date chrono.Date, numberNull null.Val[decimal.Decimal], dateNull null.Val[chrono.Date], numberNonReq omit.Val[decimal.Decimal], dateNonReq omit.Val[chrono.Date]) (TestTypeOverridesResponse, *http.Response, error) {
+func (_c Client) TestTypeOverrides(ctx context.Context, body *Primitives, number decimal.Decimal, date chrono.Date, numberNull null.Val[decimal.Decimal], dateNull null.Val[chrono.Date], numberNonReq omit.Val[decimal.Decimal], dateNonReq omit.Val[chrono.Date]) (TestTypeOverridesResponse, *http.Response, error) {
+	baseURL := _c.url
 	_urlStr := strings.TrimSuffix(baseURL.ToURL(), "/") + `/test/type_overrides`
 	_req, _err := http.NewRequestWithContext(ctx, http.MethodGet, _urlStr, nil)
 	if _err != nil {
@@ -310,7 +374,8 @@ func (_c Client) TestTypeOverrides(ctx context.Context, baseURL BaseURLBuilder, 
 }
 
 // TestUnknownBodyType post /test/unknown/body/type
-func (_c Client) TestUnknownBodyType(ctx context.Context, baseURL BaseURLBuilder, body io.ReadCloser) (TestUnknownBodyTypeResponse, *http.Response, error) {
+func (_c Client) TestUnknownBodyType(ctx context.Context, body io.ReadCloser) (TestUnknownBodyTypeResponse, *http.Response, error) {
+	baseURL := _c.url
 	_urlStr := strings.TrimSuffix(baseURL.ToURL(), "/") + `/test/unknown/body/type`
 	_req, _err := http.NewRequestWithContext(ctx, http.MethodPost, _urlStr, nil)
 	if _err != nil {
@@ -343,7 +408,8 @@ func (_c Client) TestUnknownBodyType(ctx context.Context, baseURL BaseURLBuilder
 // Retrieves a user with a long description that spans multiple lines so
 // that we can see that both wrapping and long-line support is not
 // bleeding over the sacred 80 char limit.
-func (_c Client) GetUser(ctx context.Context, baseURL BaseURLBuilder, id string, paramComponent string, validStr omitnull.Val[GetUserGetValidStrParam], reqValidStr null.Val[GetUserGetReqValidStrParam], validInt omit.Val[int], reqValidInt int, validNum omit.Val[float64], reqValidNum float64, validBool omit.Val[bool], reqValidBool bool, reqStrFormat uuid.UUID, dateTime chrono.DateTime, date chrono.Date, timeVal chrono.Time, durationVal time.Duration, arrayPrimExplode omit.Val[GetUserGetArrayPrimExplodeParam], arrayPrimFlat GetUserGetArrayPrimFlatParam, arrayPrimIntExplode omit.Val[GetUserGetArrayPrimIntExplodeParam], arrayPrimIntFlat GetUserGetArrayPrimIntFlatParam, arrayEnumExplode omit.Val[GetUserGetArrayEnumExplodeParam], arrayEnumFlat GetUserGetArrayEnumFlatParam) (GetUserResponse, *http.Response, error) {
+func (_c Client) GetUser(ctx context.Context, id string, paramComponent string, validStr omitnull.Val[GetUserGetValidStrParam], reqValidStr null.Val[GetUserGetReqValidStrParam], validInt omit.Val[int], reqValidInt int, validNum omit.Val[float64], reqValidNum float64, validBool omit.Val[bool], reqValidBool bool, reqStrFormat uuid.UUID, dateTime chrono.DateTime, date chrono.Date, timeVal chrono.Time, durationVal time.Duration, arrayPrimExplode omit.Val[GetUserGetArrayPrimExplodeParam], arrayPrimFlat GetUserGetArrayPrimFlatParam, arrayPrimIntExplode omit.Val[GetUserGetArrayPrimIntExplodeParam], arrayPrimIntFlat GetUserGetArrayPrimIntFlatParam, arrayEnumExplode omit.Val[GetUserGetArrayEnumExplodeParam], arrayEnumFlat GetUserGetArrayEnumFlatParam) (GetUserResponse, *http.Response, error) {
+	baseURL := _c.url
 	_urlStr := strings.TrimSuffix(baseURL.ToURL(), "/") + `/users/{id}`
 	_urlStr = strings.Replace(_urlStr, `{id}`, fmt.Sprintf("%v", id), 1)
 	_req, _err := http.NewRequestWithContext(ctx, http.MethodGet, _urlStr, nil)
@@ -431,7 +497,8 @@ func (_c Client) GetUser(ctx context.Context, baseURL BaseURLBuilder, id string,
 // SetUser post /users/{id}
 //
 // Sets a user
-func (_c Client) SetUser(ctx context.Context, baseURL BaseURLBuilder, body *Primitives, id string, paramComponent string) (SetUserResponse, *http.Response, error) {
+func (_c Client) SetUser(ctx context.Context, body *Primitives, id string, paramComponent string) (SetUserResponse, *http.Response, error) {
+	baseURL := _c.url
 	_urlStr := strings.TrimSuffix(baseURL.ToURL(), "/") + `/users/{id}`
 	_urlStr = strings.Replace(_urlStr, `{id}`, fmt.Sprintf("%v", id), 1)
 	_req, _err := http.NewRequestWithContext(ctx, http.MethodPost, _urlStr, nil)